# backend/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "My API",
    "description": "Description",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/api/product/": {
      "get": {
        "tags": [
          "product"
        ],
        "summary": "Retrieve all products from the database",
        "description": "This endpoint fetches a list of all products stored in the database.",
        "responses": {
          "200": {
            "description": "List of all products.",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "number",
                    "example": 1
                  },
                  "name": {
                    "type": "string",
                    "example": "Product Name"
                  },
                  "description": {
                    "type": "string",
                    "example": "Product Description"
                  },
                  "price": {
                    "type": "number",
                    "example": 99.99
                  },
                  "contactPersonId": {
                    "type": "number",
                    "example": 1
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Error message describing the issue."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/api/product/name/{name}": {
      "get": {
        "tags": [
          "product"
        ],
        "summary": "Retrieve a product by its name",
        "description": "This endpoint fetches a list of products that match the specified name in the database.",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The name of the product to search for."
          }
        ],
        "responses": {
          "200": {
            "description": "List of products with the specified name.",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "number",
                    "example": 1
                  },
                  "name": {
                    "type": "string",
                    "example": "Product Name"
                  },
                  "description": {
                    "type": "string",
                    "example": "Product Description"
                  },
                  "price": {
                    "type": "number",
                    "example": 99.99
                  },
                  "contactPersonId": {
                    "type": "number",
                    "example": 1
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "Product not found with the specified name.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "No such product name: {name}"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Error message describing the issue."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/api/product/cp-id/{contactPersonId}": {
      "get": {
        "tags": [
          "product"
        ],
        "summary": "Retrieve products by contact person ID",
        "description": "This endpoint fetches a list of products associated with a specific contact person ID.",
        "parameters": [
          {
            "name": "contactPersonId",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "The contact person ID to filter products."
          }
        ],
        "responses": {
          "200": {
            "description": "List of products associated with the specified contact person ID.",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "number",
                    "example": 1
                  },
                  "name": {
                    "type": "string",
                    "example": "Product Name"
                  },
                  "description": {
                    "type": "string",
                    "example": "Product Description"
                  },
                  "price": {
                    "type": "number",
                    "example": 99.99
                  },
                  "contactPersonId": {
                    "type": "number",
                    "example": 1
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "No products found for the specified contact person ID.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "No such product contact person id: {contactPersonId}"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Error message describing the issue."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/api/product/{id}": {
      "get": {
        "tags": [
          "product"
        ],
        "summary": "Retrieve a product by its ID",
        "description": "This endpoint fetches a product by its unique ID from the database.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "The ID of the product to fetch."
          }
        ],
        "responses": {
          "200": {
            "description": "Product details by ID.",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 1
                },
                "name": {
                  "type": "string",
                  "example": "Product Name"
                },
                "description": {
                  "type": "string",
                  "example": "Product Description"
                },
                "price": {
                  "type": "number",
                  "example": 99.99
                },
                "contactPersonId": {
                  "type": "number",
                  "example": 1
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "Product not found with the specified ID.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "No such product with id: {id}"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Error message describing the issue."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
...
  }
}